+++ ItemExtensions/Patches/Drops/FishingRod.cs
@@ -1,6 +1,7 @@
 using System.Reflection.Emit;
 using HarmonyLib;
 using ItemExtensions.Additions;
+using ItemExtensions.Models.Items;
 using StardewModdingAPI;
 using StardewValley;
 using StardewValley.Internal;
@@ -49,6 +50,9 @@ internal class FishingRodPatches
 #if DEBUG
             Log($"Checking entry {entry}...");
 #endif
+            if (RodMatch(data, __instance) == false)
+                continue;
+            
             if (Sorter.GetItem(data, context, out var item) == false)
                 continue;
 
@@ -56,5 +60,63 @@ internal class FishingRodPatches
             
             Log($"Added treasure reward from entry {entry} ({item.QualifiedItemId})");
         }
+    }
+
+    /// <summary>
+    /// Checks for rod conditions.
+    /// </summary>
+    /// <param name="data">The spawn data.</param>
+    /// <param name="rod">The rod to check</param>
+    /// <returns>Whether all conditions match.</returns>
+    private static bool RodMatch(TreasureData data, FishingRod rod)
+    {
+        if (rod is null)
+            return false;
+        
+        try
+        {
+            //rods
+            if (data.Rod.Any() && data.Rod.Contains(rod.ItemId) == false)
+                return false;
+            //bait
+            if (rod.GetBait() != null)
+            {
+                if (data.Bait.Any() && data.Bait.Contains(rod.GetBait().ItemId) == false)
+                    return false;
+            }
+            //tackle
+            if (rod.GetTackle() != null && rod.GetTackle().Any() && data.Tackle.Any())
+            {
+                //if all tackle must be in list
+                if (data.RequireAllTackle)
+                {
+                    if (rod.GetTackle().TrueForAll(i => data.Tackle.Contains(i.ItemId)) == false)
+                        return false;
+                }
+                else if (rod.GetTackle().Any(i => data.Tackle.Contains(i.ItemId)) == false)
+                    return false;
+            }
+            //attachment limits
+            if (data.MinAttachments > 0)
+            {
+                if (rod.AttachmentSlotsCount < data.MinAttachments)
+                    return false;
+            }
+            if (data.MaxAttachments > -1)
+            {
+                if (rod.AttachmentSlotsCount > data.MinAttachments)
+                    return false;
+            }
+            //bobber
+            if (data.Bobber >= 0 && data.Bobber != rod.getBobberStyle(rod.getLastFarmerToUse()))
+                return false;
+        }
+        catch (Exception e)
+        {
+            Log($"Error: {e}.\nCheck will be treated as false.", LogLevel.Warn);
+            return false;
+        }
+
+        return true;
     }
 }